Cleaned up components and made reusable components
