+++ src/store/modules/events.js
@@ -30,35 +30,37 @@ const actions = {
 /* ===== MUTATIONS ===== */
 const mutations = {
   RECEIVE_EVENTS(state, {data}) {
-    const days = [
-      'Sunday',
-      'Monday',
-      'Tuesday',
-      'Wednesday',
-      'Thursday',
-      'Friday',
-      'Saturday'
-    ]
-    data.forEach(singleEvent => {
+    // const days = [
+    //   'Sunday',
+    //   'Monday',
+    //   'Tuesday',
+    //   'Wednesday',
+    //   'Thursday',
+    //   'Friday',
+    //   'Saturday'
+    // ]
+    for (let event of data) {
       state.events.push({
-        title: singleEvent.name,
-        date:
-          new Date(singleEvent.time).toLocaleDateString('en-GB') +
-          ' (' +
-          days[new Date(singleEvent.time).getDay()] +
-          ')' +
+        title: event.name,
+        milli: event.time,
+        description: event.description,
+        status: event.status,
+        url: event.event_url,
+        info: {
+          date:
+          new Date(event.time).toLocaleDateString('en-GB') +
+          // ' (' +
+          // days[new Date(event.time).getDay()] +
+          // ')' +
           ' @ ' +
-          new Date(singleEvent.time)
+          new Date(event.time)
             .toLocaleTimeString('en-GB')
             .substring(0, 5),
-        milli: singleEvent.time,
-        description: singleEvent.description,
-        location: singleEvent.venue.name,
-        attending: singleEvent.yes_rsvp_count,
-        status: singleEvent.status,
-        url: singleEvent.event_url
+          location: event.venue.name,
+          attending: event.yes_rsvp_count
+        }
       })
-    })
+    }
   }
 }
