+++ src/store/modules/tweets.js
@@ -14,12 +14,22 @@ const getters = {
 /* ===== ACTIONS ===== */
 const actions = {
   // FETCH EVENTS FROM MEETUP API
-  async FETCH_FEED({commit}) {
+  async FETCH_USER_TWEETS({commit}) {
     try {
       const response = await axios.get(
-        'https://37vqk7ncii.execute-api.us-east-1.amazonaws.com/events'
+        'https://hpteg5u9ub.execute-api.us-east-1.amazonaws.com/tweets'
       )
-      commit('RECEIVE_FEED', {data: response.data.results})
+      commit('RECEIVE_TWEETS', {data: response.data})
+    } catch (error) {
+      console.log(error)
+    }
+  },
+  async FETCH_HASHTAGS({commit}) {
+    try {
+      const response = await axios.get(
+        'https://qkwbqp2lbd.execute-api.us-east-1.amazonaws.com/hashtags'
+      )
+      commit('RECEIVE_TWEETS', {data: response.data.statuses})
     } catch (error) {
       console.log(error)
     }
@@ -28,20 +38,25 @@ const actions = {
 
 /* ===== MUTATIONS ===== */
 const mutations = {
-  RECEIVE_FEED(state, {data}) {
-    data.forEach(singleEvent => {
-      state.feed.push({
-        title: singleEvent.name,
-        date: new Date(singleEvent.time).toLocaleDateString('en-GB'),
-        milli: singleEvent.time,
-        description: singleEvent.description,
-        location: singleEvent.venue.name,
-        url: singleEvent.event_url,
-        status: singleEvent.status,
-        type: 'event',
-        where: 'meetup'
+  RECEIVE_TWEETS(state, {data}) {
+    // const tempTweets = []
+    for (let tweet of data) {
+      let img = tweet.user.profile_image_url_https
+      let imgHq = img.substring(0, img.lastIndexOf('_'))
+      state.tweets.push({
+        title: tweet.user.screen_name,
+        description: tweet.text,
+        id: tweet.id,
+        milli: new Date(tweet.created_at).getTime(),
+        profile_img: imgHq,
+        media: tweet.entities.media,
+        info: {
+          date: new Date(tweet.created_at).toLocaleDateString('en-GB'),
+          retweets: tweet.retweet_count,
+          favorites: tweet.favorite_count
+        }
       })
-    })
+    }
   }
 }
